+++ backend/runtimes/node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @cat-factory/node-server
 
+## 0.106.5
+
+### Patch Changes
+
+- Updated dependencies [edfd2f8]
+- Updated dependencies [d675cc5]
+  - @cat-factory/orchestration@0.123.5
+  - @cat-factory/server@0.137.4
+
 ## 0.106.4
 
 ### Patch Changes